+++ intervals/parser.py
@@ -1,4 +1,3 @@
-from collections import Iterable
 from .exc import IntervalException
 try:
     string_types = basestring,  # Python 2
@@ -9,30 +8,21 @@ except NameError:
 strip = lambda a: a.strip()
 
 
-class IntervalParser(object):
-    def parse_object(self, obj):
-        return obj.lower, obj.upper, obj.lower_inc, obj.upper_inc
-
+class IntervalStringParser(object):
     def parse_string(self, value):
         if ',' not in value:
             return self.parse_hyphen_range(value)
         else:
             return self.parse_bounded_range(value)
 
-    def parse_sequence(self, seq):
-        lower, upper = seq
-        if isinstance(seq, tuple):
-            return lower, upper, False, False
-        else:
-            return lower, upper, True, True
-
-    def parse_single_value(self, value):
-        return value, value, True, True
-
     def parse_bounded_range(self, value):
         values = value.strip()[1:-1].split(',')
         lower, upper = map(strip, values)
-        return lower, upper, value[0] == '[', value[-1] == ']'
+        return (
+            [lower, upper],
+            value[0] == '[',
+            value[-1] == ']'
+        )
 
     def parse_hyphen_range(self, value):
         """
@@ -64,12 +54,25 @@ class IntervalParser(object):
                     values_copy.append(value)
             lower, upper = values_copy
 
-        return lower, upper, True, True
+        return [lower, upper], True, True
+
+
+class IntervalParser(object):
+    def parse_object(self, obj):
+        return obj.lower, obj.upper, obj.lower_inc, obj.upper_inc
+
+    def parse_sequence(self, seq):
+        lower, upper = seq
+        if isinstance(seq, tuple):
+            return lower, upper, False, False
+        else:
+            return lower, upper, True, True
+
+    def parse_single_value(self, value):
+        return value, value, True, True
 
     def __call__(self, bounds, lower_inc=None, upper_inc=None):
-        if isinstance(bounds, string_types):
-            values = self.parse_string(bounds)
-        elif isinstance(bounds, Iterable):
+        if isinstance(bounds, (list, tuple)):
             values = self.parse_sequence(bounds)
         elif hasattr(bounds, 'lower') and hasattr(bounds, 'upper'):
             values = self.parse_object(bounds)