+++ tests/interval/test_operators.py
@@ -1,7 +1,8 @@
+from datetime import date
+
 from pytest import mark
 from infinity import inf
-from intervals import IntInterval
-from intervals import FloatInterval
+from intervals import DateInterval, FloatInterval, IntInterval
 
 
 class TestComparisonOperators(object):
@@ -11,7 +12,13 @@ class TestComparisonOperators(object):
         (IntInterval([inf, inf]) == inf, True),
         (IntInterval([3, 3]) == 3, True),
         (IntInterval([3, 3]) == 5, False),
-        (IntInterval('(,)') == None, False)
+        (IntInterval([3, 3]) == 'something', False),
+        (
+            IntInterval([3, 3]) ==
+            DateInterval([date(2011, 1, 1), date(2011, 1, 1)]),
+            False
+        ),
+        (IntInterval.from_string('(,)') == None, False)
     ))
     def test_eq_operator(self, comparison, result):
         assert comparison is result
@@ -22,7 +29,7 @@ class TestComparisonOperators(object):
         (IntInterval([inf, inf]) != inf, False),
         (IntInterval([3, 3]) != 3, False),
         (IntInterval([3, 3]) != 5, True),
-        (IntInterval('(,)') != None, True)
+        (IntInterval.from_string('(,)') != None, True)
     ))
     def test_ne_operator(self, comparison, result):
         assert comparison is result
@@ -82,7 +89,7 @@ class TestComparisonOperators(object):
     @mark.parametrize('value', (
         IntInterval([0, 2]),
         2,
-        '[-1, 1]',
+        [-1, 1],
     ))
     def test_contains_operator_for_non_inclusive_interval(self, value):
         assert value not in IntInterval((-1, 2))
@@ -90,8 +97,16 @@ class TestComparisonOperators(object):
     @mark.parametrize(('interval1', 'interval2', 'expected'), (
         (IntInterval((0, 2)), IntInterval((0, 2)), True),
         (IntInterval([0, 2]), IntInterval([0, 2]), True),
-        (IntInterval('[0, 2)'), IntInterval('[0, 2)'), True),
-        (IntInterval('(0, 2]'), IntInterval('(0, 2]'), True),
+        (
+            IntInterval.from_string('[0, 2)'),
+            IntInterval.from_string('[0, 2)'),
+            True
+        ),
+        (
+            IntInterval.from_string('(0, 2]'),
+            IntInterval.from_string('(0, 2]'),
+            True
+        ),
         (IntInterval((0, 2)), IntInterval((1, 2)), False),
         (IntInterval((0, 2)), IntInterval((0, 1)), False),
         (IntInterval((0, 2)), IntInterval([0, 1]), False),
@@ -112,12 +127,15 @@ class TestComparisonOperators(object):
 
 class TestDiscreteRangeComparison(object):
     @mark.parametrize(('interval', 'interval2'), (
-        ([1, 3], '[1, 4)'),
+        ('[1, 3]', '[1, 4)'),
         ('(1, 5]', '[2, 5]'),
         ('(1, 6)', '[2, 5]'),
     ))
     def test_eq_operator(self, interval, interval2):
-        assert IntInterval(interval) == IntInterval(interval2)
+        assert (
+            IntInterval.from_string(interval) ==
+            IntInterval.from_string(interval2)
+        )
 
 
 class TestBinaryOperators(object):
@@ -125,8 +143,6 @@ class TestBinaryOperators(object):
         ((2, 3), (3, 4), (3, 3)),
         ((2, 3), [3, 4], (3, 3)),
         ((2, 5), (3, 10), (3, 5)),
-        ('(2, 3]', '[3, 4)', [3, 3]),
-        ('(2, 10]', '[3, 40]', [3, 10]),
         ((2, 10), (3, 8), (3, 8)),
         ((1, 2), [1, 2], (1, 2)),
         ([1, 2], (1, 2), (1, 2)),
@@ -137,11 +153,41 @@ class TestBinaryOperators(object):
             IntInterval(result)
         )
 
+    @mark.parametrize(('interval1', 'interval2', 'result'), (
+        ('(2, 3]', '[3, 4)', [3, 3]),
+        ('(2, 10]', '[3, 40]', [3, 10]),
+    ))
+    def test_and_operator_with_half_open_intervals(
+        self,
+        interval1,
+        interval2,
+        result
+    ):
+        assert (
+            IntInterval.from_string(interval1) &
+            IntInterval.from_string(interval2) ==
+            IntInterval(result)
+        )
+
     @mark.parametrize(('interval1', 'interval2', 'empty'), (
         ((2, 3), (3, 4), True),
         ((2, 3), [3, 4], True),
         ([2, 3], (3, 4), True),
-        ('(2, 3]', '[3, 4)', False),
     ))
     def test_and_operator_for_empty_results(self, interval1, interval2, empty):
         assert (IntInterval(interval1) & IntInterval(interval2)).empty == empty
+
+    @mark.parametrize(('interval1', 'interval2', 'empty'), (
+        ('(2, 3]', '[3, 4)', False),
+    ))
+    def test_and_operator_for_half_open_intervals_with_empty_results(
+        self,
+        interval1,
+        interval2,
+        empty
+    ):
+        assert (
+            IntInterval.from_string(interval1) &
+            IntInterval.from_string(interval2)
+        ).empty == empty
+